+++ android/java/com/sometrik/framework/NativeSurface.java
@@ -65,7 +65,7 @@ public class NativeSurface extends SurfaceView implements NativeCommandHandler {
     //TODO
   }
   @Override
-  public void addOption(long optionId, String text) {
+  public void addOption(int optionId, String text) {
     System.out.println("Native Surface couldn't handle command");
   }
